remove tagslist in favor of commontags
